+++ tests/collect_memory/memory_collect_test.c
@@ -6,7 +6,6 @@
  * Author:      Podola Radim, xpodol06@stud.fit.vutbr.cz
  * Description: Testing file for injected malloc.so library.
 
-TODO: Test for all allocation functions, use assert?
  */
 #include <stdio.h>
 #include <stdlib.h>
@@ -21,16 +20,23 @@ void fun() {
 }
 
 
-int main(){
+int main(int argc, char** argv){
 
 	int *n, *m, res;
+
+	int repeats = 1;
+	if (argc == 2) {
+        repeats = atoi(argv[1]);
+	}
 
 	n = (int*)malloc(sizeof(int));
 	*n = 5;
 	free(n);
 
-	for (int i = 0; i < 1000000; ++i);
-	fun();
+    while(repeats--) {
+        for (int i = 0; i < 1000000; ++i);
+        fun();
+	}
 
 	m = (int*)malloc(sizeof(int));
 	*m = 5;